Breakout the subplot widgets and re-add logic to set shared size policy.
